Proof of concept for feeding a browser with Server side events
Reflections:
* While this somewhat hacky server works together with a single
local file, it's very limiting. Can't have more than one file
being the most critical if I want to use wasm.
* I will want some kind of event list for this, rather than parsing
the context list, which is what I want for logging as well,
so the whole idea of storing all contexts is probably not as useful
as I first though. The context object themselves is a decent way
to wrap event generation in a way where you can't create invalid
event sequences, but they should probably be short lived.
* The whole one-client-at-a-time concept probably doesn't matter,
or rather, it's not really working currently, since a second client
can connect and block the queue if the first client reloads the page.
A normal server on a random port on localhost should be more than
safe enough, even if i just use http.server.