Refine server
Switched from server imitation to using http.server library.

While http.server is documented to not be suitable for production,
that shouldn't matter as long as we are using localhost, even more
so with unpredictable port numbers.

The page is now served through the server as well, rather than being
fed to webbrowser from file directly, this eliminates the need to
add cross site scripting exceptions in the headers.